fetchers/channel_fetcher.py
"""
Channel and video discovery from YouTube Data API.

Handles:
- Channel metadata fetching
- Video list retrieval
- Video metadata fetching
"""

import os
from typing import Any
from googleapiclient.discovery import build

from database.db_manager import DatabaseManager
from utils.logger import get_logger

logger = get_logger(__name__)


class ChannelFetcher:
    """
    Fetches channel and video metadata from YouTube Data API v3.
    """

    def __init__(self, db_manager: DatabaseManager):
        """
        Initialize channel fetcher.

        Args:
            db_manager: Database manager instance.
        """
        self.db_manager = db_manager
        self.youtube = build(
            "youtube",
            "v3",
            developerKey=os.getenv("YOUTUBE_API_KEY")
        )

    async def discover_channels(self, channel_ids: list[str]) -> None:
        """
        Discover and process all configured channels.

        For each channel:
        1. Fetch channel metadata
        2. Fetch all video IDs
        3. Fetch video metadata
        4. Initialize status records

        Args:
            channel_ids: List of YouTube channel IDs to process.
        """
        logger.info(f"🔄 Discovering {len(channel_ids)} channels")

        for channel_id in channel_ids:
            await self._process_channel(channel_id)

        logger.info("✅ Channel discovery completed")

    async def _process_channel(self, channel_id: str) -> None:
        """
        Process a single channel.

        Args:
            channel_id: YouTube channel ID.
        """
        # TODO: Implement channel processing
        # 1. Check if channel exists in database
        # 2. If not, fetch channel name from API and insert
        # 3. Fetch all video IDs from channel
        # 4. For each video, fetch metadata and insert if new
        # 5. Initialize Status records for new videos

        logger.info(f"🔄 Processing channel: {channel_id}")
        pass

    def _fetch_channel_metadata(self, channel_id: str) -> dict[str, Any]:
        """
        Fetch channel metadata from YouTube API.

        Args:
            channel_id: YouTube channel ID.

        Returns:
            Channel metadata dictionary.
        """
        # TODO: Implement API call to fetch channel name
        # API endpoint: youtube.channels().list(part="snippet", id=channel_id)
        pass

    def _fetch_channel_videos(self, channel_id: str) -> list[str]:
        """
        Fetch all video IDs from a channel.

        Args:
            channel_id: YouTube channel ID.

        Returns:
            List of video IDs.
        """
        # TODO: Implement API call to fetch video list
        # 1. Get uploads playlist ID from channel
        # 2. Paginate through playlist items
        # API endpoint: youtube.playlistItems().list()
        pass

    def _fetch_video_metadata(self, video_id: str) -> dict[str, Any]:
        """
        Fetch video metadata from YouTube API.

        Args:
            video_id: YouTube video ID.

        Returns:
            Video metadata dictionary with fields:
            - video_title
            - published_date
            - duration_seconds
            - view_count
            - like_count
        """
        # TODO: Implement API call to fetch video details
        # API endpoint: youtube.videos().list(part="snippet,contentDetails,statistics", id=video_id)
        # Parse duration from ISO 8601 format (e.g., "PT4M13S") to seconds
        pass
