utils/gemini_client.py
"""
Gemini API client wrapper.

Provides simplified interface for Gemini API calls with error handling.
"""

import os
from typing import Any, Optional
import google.generativeai as genai

import config
from utils.logger import get_logger

logger = get_logger(__name__)


class GeminiClient:
    """
    Wrapper for Gemini API calls.

    Handles:
    - API configuration
    - Request execution
    - Error handling
    - Response parsing
    """

    def __init__(self, model_code: str):
        """
        Initialize Gemini client.

        Args:
            model_code: Gemini model code (e.g., 'gemini-2.5-flash-lite').
        """
        genai.configure(api_key=os.getenv("GEMINI_API_KEY"))
        self.model_code = model_code
        self.model = genai.GenerativeModel(model_code)

    async def generate_content(
        self,
        prompt: str,
        response_mime_type: str = "application/json"
    ) -> dict[str, Any]:
        """
        Generate content from prompt.

        Args:
            prompt: Input prompt string.
            response_mime_type: Expected response MIME type.

        Returns:
            Dictionary with:
            - response_text: Generated text
            - input_tokens: Number of input tokens
            - output_tokens: Number of output tokens
        """
        # TODO: Implement Gemini API call
        # 1. Configure generation with JSON response type
        # 2. Call model.generate_content()
        # 3. Extract response text
        # 4. Get token counts from usage_metadata
        # 5. Return structured response

        pass

    async def generate_embedding(self, text: str) -> list[float]:
        """
        Generate embedding for text.

        Args:
            text: Input text to embed.

        Returns:
            Embedding vector as list of floats.
        """
        # TODO: Implement embedding generation
        # 1. Use genai.embed_content() with embedding model
        # 2. Specify dimension size from config
        # 3. Return embedding vector

        pass
